requirements.txt
pandas>=2.2,<2.4
numpy>=2.1,<2.3
scikit-learn>=1.5,<1.7
requests>=2.31,<3
python-dotenv>=1.0,<2
openai>=1.30,<2
rich>=13.7,<14
joblib>=1.3,<2
pytest>=8,<9
